[internal] move state into its own module
