+++ src/keyboard.rs
@@ -10,67 +10,12 @@ use crate::hardware::{Dir, Encoder};
 use crate::keycode::Keycode;
 
 mod matrix;
+mod state;
 mod usb;
 
 use matrix::Matrix;
+use state::State;
 use usb::Usb;
-
-struct Layout<'a> {
-    layout: &'a [&'a [&'a [Keycode]]],
-}
-
-impl<'a> Layout<'a> {
-    fn new(layout: &'a [&[&[Keycode]]]) -> Self {
-        Self { layout }
-    }
-}
-
-struct State<'a, const NUM_OF_COLS: usize, const NUM_OF_ROWS: usize> {
-    layout: Layout<'a>,
-    layer: usize,
-    override_keys: [[Option<u8>; NUM_OF_COLS]; NUM_OF_ROWS],
-}
-
-impl<'a, const NUM_OF_COLS: usize, const NUM_OF_ROWS: usize> State<'a, NUM_OF_COLS, NUM_OF_ROWS> {
-    fn new(layout: &'a [&[&[Keycode]]]) -> Self {
-        Self {
-            layout: Layout::new(layout),
-            layer: 0,
-            override_keys: [[None; NUM_OF_COLS]; NUM_OF_ROWS],
-        }
-    }
-
-    fn get_key(&self, row: usize, col: usize) -> Keycode {
-        if let Some(layer) = self.override_keys[row][col] {
-            self.layout.layout[layer as usize][row][col]
-        } else {
-            self.layout.layout[self.layer][row][col]
-        }
-    }
-
-    // handles special press actions
-    fn on_press(&mut self, keycode: Keycode, row: usize, col: usize) {
-        match keycode {
-            Keycode::KC_MO(layer) => {
-                self.override_keys[row][col] = Some(self.layer.clone() as u8);
-                self.layer = layer;
-            }
-            Keycode::KC_LAYER(layer) => self.layer = layer,
-            _ => {}
-        }
-    }
-
-    // handles special release actions
-    fn on_release(&mut self, keycode: Keycode, row: usize, col: usize) {
-        match keycode {
-            Keycode::KC_MO(_) => {
-                self.layer = self.override_keys[row][col].unwrap() as usize;
-                self.override_keys[row][col] = None;
-            }
-            _ => {}
-        }
-    }
-}
 
 pub struct Keyboard<'a, const NUM_OF_COLS: usize, const NUM_OF_ROWS: usize> {
     state: State<'a, NUM_OF_COLS, NUM_OF_ROWS>,
